refactor: move cloze hook to feature
